requirements.txt
matplotlib==3.5.2
numpy==1.23.1
opencv_python==4.6.0.66
Pillow==9.2.0
scikit_image==0.19.3
skimage
tqdm==4.64.0
treepoem==3.16.0
wget==3.2
